Play movement is physics based

+++ src/main/java/com/JEngine/PrimitiveTypes/Position/Vector2.java
@@ -29,7 +29,7 @@ public class Vector2 {
     }
     public Vector2 multiply(Vector2 v)
     {
-        return new Vector2((float)(x * v.x), (float)(y * v.y));
+        return new Vector2((x * v.x), (y * v.y));
     }
     public Vector2 add(Vector2 v)
     {
